feat: Added Splashscreen, Facebook Login and HomeFragment
Noticed that the application takes a while to load and so I added a splashscreen to hide this.

LoginActivity is implemented for only facebook for the time being. Other login options will be added later.

The HomeFragment will be the first view to be displayed in the tabbedview and this will show active discounts and coupons based on user preferences. I added demo pictures to the app for the discounts.